refactor: 회원 패키지 정리
