+++ src/main/java/com/alzzaipo/member/application/port/in/account/local/CheckLocalAccountIdAvailableQuery.java
@@ -2,7 +2,7 @@ package com.alzzaipo.member.application.port.in.account.local;
 
 import com.alzzaipo.member.domain.account.local.LocalAccountId;
 
-public interface CheckLocalAccountIdAvailabilityQuery {
+public interface CheckLocalAccountIdAvailableQuery {
 
     boolean checkAccountIdAvailable(LocalAccountId localAccountId);
 }
