+++ src/main/java/com/alzzaipo/member/application/port/in/account/social/KakaoLoginUseCase.java
@@ -1,4 +1,4 @@
-package com.alzzaipo.member.application.port.in.oauth;
+package com.alzzaipo.member.application.port.in.account.social;
 
 import com.alzzaipo.member.application.port.in.dto.AuthorizationCode;
 import com.alzzaipo.member.application.port.in.dto.LoginResult;
