+++ src/main/java/com/alzzaipo/member/application/port/in/account/social/LinkKakaoAccountUseCase.java
@@ -1,4 +1,4 @@
-package com.alzzaipo.member.application.port.in.oauth;
+package com.alzzaipo.member.application.port.in.account.social;
 
 import com.alzzaipo.common.Uid;
 import com.alzzaipo.member.application.port.in.dto.AuthorizationCode;
