+++ src/main/java/com/alzzaipo/member/application/service/account/SocialAccountService.java
@@ -1,12 +1,17 @@
-package com.alzzaipo.member.application.service.account.social;
+package com.alzzaipo.member.application.service.account;
 
 import com.alzzaipo.common.LoginType;
 import com.alzzaipo.common.Uid;
 import com.alzzaipo.common.exception.CustomException;
+import com.alzzaipo.common.token.TokenUtil;
+import com.alzzaipo.common.token.application.port.out.SaveRefreshTokenPort;
+import com.alzzaipo.common.token.domain.TokenInfo;
 import com.alzzaipo.member.application.port.in.account.social.UnlinkSocialAccountUseCase;
 import com.alzzaipo.member.application.port.in.dto.AuthorizationCode;
+import com.alzzaipo.member.application.port.in.dto.LoginResult;
 import com.alzzaipo.member.application.port.in.dto.UnlinkSocialAccountCommand;
-import com.alzzaipo.member.application.port.in.oauth.LinkKakaoAccountUseCase;
+import com.alzzaipo.member.application.port.in.account.social.KakaoLoginUseCase;
+import com.alzzaipo.member.application.port.in.account.social.LinkKakaoAccountUseCase;
 import com.alzzaipo.member.application.port.out.account.local.FindLocalAccountByMemberUidPort;
 import com.alzzaipo.member.application.port.out.account.social.DeleteSocialAccountPort;
 import com.alzzaipo.member.application.port.out.account.social.FindSocialAccountPort;
@@ -14,9 +19,11 @@ import com.alzzaipo.member.application.port.out.account.social.RegisterSocialAcc
 import com.alzzaipo.member.application.port.out.dto.AccessToken;
 import com.alzzaipo.member.application.port.out.dto.FindSocialAccountCommand;
 import com.alzzaipo.member.application.port.out.dto.UserProfile;
-import com.alzzaipo.member.application.port.out.oauth.ExchangeKakaoAccessTokenPort;
-import com.alzzaipo.member.application.port.out.oauth.FetchKakaoUserProfilePort;
+import com.alzzaipo.member.application.port.out.account.social.ExchangeKakaoAccessTokenPort;
+import com.alzzaipo.member.application.port.out.account.social.FetchKakaoUserProfilePort;
+import com.alzzaipo.member.application.port.out.member.RegisterMemberPort;
 import com.alzzaipo.member.domain.account.social.SocialAccount;
+import com.alzzaipo.member.domain.member.Member;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
@@ -24,7 +31,8 @@ import org.springframework.stereotype.Service;
 @Service
 @RequiredArgsConstructor
 public class SocialAccountService implements LinkKakaoAccountUseCase,
-    UnlinkSocialAccountUseCase {
+    UnlinkSocialAccountUseCase,
+    KakaoLoginUseCase {
 
     private final FindLocalAccountByMemberUidPort findLocalAccountByMemberUidPort;
     private final ExchangeKakaoAccessTokenPort exchangeKakaoAccessTokenPort;
@@ -32,6 +40,9 @@ public class SocialAccountService implements LinkKakaoAccountUseCase,
     private final FindSocialAccountPort findSocialAccountPort;
     private final RegisterSocialAccountPort registerSocialAccountPort;
     private final DeleteSocialAccountPort deleteSocialAccountUsePort;
+    private final SaveRefreshTokenPort saveRefreshTokenPort;
+    private final RegisterMemberPort registerMemberPort;
+
 
     @Override
     public void linkKakaoAccount(Uid memberUID, AuthorizationCode authorizationCode) {
@@ -63,6 +74,39 @@ public class SocialAccountService implements LinkKakaoAccountUseCase,
             .orElseThrow(() -> new CustomException(HttpStatus.FORBIDDEN, "로컬 계정에서만 가능합니다"));
 
         deleteSocialAccountUsePort.deleteSocialAccount(command.getMemberUID(), command.getLoginType());
+    }
+
+    @Override
+    public LoginResult handleLogin(AuthorizationCode authorizationCode) {
+        try {
+            AccessToken kakaoAccessToken
+                = exchangeKakaoAccessTokenPort.exchangeKakaoAccessToken(authorizationCode);
+
+            UserProfile kakaoUserProfile
+                = fetchKakaoUserProfilePort.fetchKakaoUserProfile(kakaoAccessToken);
+
+            FindSocialAccountCommand command = new FindSocialAccountCommand(LoginType.KAKAO, kakaoUserProfile.getEmail());
+            SocialAccount socialAccount = findSocialAccountPort.findSocialAccount(command)
+                .orElseGet(() -> registerSocialAccount(kakaoUserProfile));
+
+            TokenInfo tokenInfo = TokenUtil.createToken(socialAccount.getMemberUID(), LoginType.KAKAO);
+
+            saveRefreshTokenPort.save(tokenInfo.getRefreshToken(), socialAccount.getMemberUID());
+
+            return new LoginResult(true, tokenInfo);
+        } catch (Exception e) {
+            return LoginResult.getFailedResult();
+        }
+    }
+
+    private SocialAccount registerSocialAccount(UserProfile userProfile) {
+        Member member = Member.build(userProfile.getNickname());
+        registerMemberPort.registerMember(member);
+
+        SocialAccount socialAccount = new SocialAccount(member.getUid(), userProfile.getEmail(), LoginType.KAKAO);
+        registerSocialAccountPort.registerSocialAccount(socialAccount);
+
+        return socialAccount;
     }
 
 }
